Fix various warnings and printing of uint64_t bench 3973806

+++ src/perft.c
@@ -1,6 +1,7 @@
 #include <time.h>
 #include <math.h>
 #include <stdio.h>
+#include <inttypes.h>
 
 #include "perft.h"
 #include "utility.h"
@@ -33,12 +34,12 @@ uint64_t splitPerft(Board *board, const int depth) {
     for (int i = 0; i < num_moves; i++) {
         MakeMove(board, moves[i]);
         const uint64_t u = perft(board, depth - 1);
-        printf("%s: %llu\n", MoveToString(moves[i]), u);
+        printf("%s: %"PRIu64"\n", MoveToString(moves[i]), u);
         nodes += u;
         *board = copy;
     }
     double elapsed = clock() - start;
-    printf("\nNodes searched: %llu\n", nodes);
-    printf("Knps: %lf\n", round(nodes / elapsed));
+    printf("\nNodes searched: %"PRIu64"\n", nodes);
+    printf("Knps: %lf\n", round((double)nodes / elapsed));
     return nodes;
 }
